move trailing slash workaround to content-page
